🐛 fix|RegionElevation|Infinite loops when plateau/ramp combine
Redo finding the closest segment to a position. Focus on fact that the segments represent a straight line and narrow them down from there. Should be faster and more accurate than using 3d distance.

Don't shift regions if the intersection is at the "a" point. Prevents infinite loop where we just switched regions and now want to not switch back. E.g., a ramp that moves lower to hit a plateau, so that we move from ramp to plateau.