🎸 feat |RegionElevation|Add nearestGroundElevation static method
Determine the ground elevation from a given position. If not within a region, falls to the terrain floor. If in a region, go to the plateau for that region. So, in general, goes to the highest plateau at a 2d location. Unless the region's bottom is above the elevation (it is floating) at which point it would go to the terrain floor.