💡 refactor|Plateau|WIP Attempt to use region behaviors for plateaus
Works okay for entry but fails when using multiple types -- entry, move within, move out. Either ignores the second move or gets stuck in a loop. Also doesn't work at all with the ruler. Better approach seems to be to treat terrain as part of the pathfinding to be done earlier.